+++ post-trade/src/test/java/com/multiassetoms/posttrade/application/SettlementServiceTest.java
@@ -196,6 +196,7 @@ class SettlementServiceTest {
                 new BigDecimal("55000.0000000000"),
                 new BigDecimal("550000"),
                 null,
+                null,
                 status,
                 capturedAt,
                 settledAt,